+++ samples/SampleWebApp/Pages/Index.cshtml.cs
@@ -1,0 +1,7 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace SampleWebApp.Pages;
+
+public class IndexModel : PageModel
+{
+}